backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.38
langchain==0.3.0
langchain-groq==0.2.0
langchain-core==0.3.0
python-dotenv==1.0.1
pydantic==2.9.2
